Revert "Revert "Merge branch 'community' into Blockade-running-missions""
This reverts commit 0d758c82e21235aa0384d6921ce0cfd3e8a5b4a2.

+++ source/OutlineShader.h
@@ -26,7 +26,7 @@ class OutlineShader {
 public:
 	static void Init();
 	
-	static void Draw(const Sprite *sprite, const Point &pos, const Point &size, const Color &color, const Point &unit = Point(0., -1.), int frame = 0);
+	static void Draw(const Sprite *sprite, const Point &pos, const Point &size, const Color &color, const Point &unit = Point(0., -1.), float frame = 0.f);
 };
